+++ assets/javascript/apps/pipeline/nodes/GetInputWidget.tsx
@@ -1,12 +1,7 @@
-import {
-  HistoryTypeWidget as HistoryTypeWidgetInner,
-  ExpandableTextWidget,
-  InputField, LlmWidget, KeywordsWidget as KeywordsWidgetInner
-} from "../widgets";
 import React from "react";
-import {concatenate, getSelectOptions} from "../utils";
-import {PropertySchema, NodeParams} from "../types/nodeParams";
+import {NodeParams, PropertySchema} from "../types/nodeParams";
 import usePipelineManagerStore from "../stores/pipelineManagerStore";
+import {getWidget} from "./widgets";
 
 
 type InputWidgetParams = {
@@ -89,147 +84,3 @@ export const getInputWidget = (params: InputWidgetParams) => {
     />
   )
 };
-
-function getWidget(name: string) {
-  switch (name) {
-    case "toggle":
-      return ToggleWidget
-    case "float":
-      return FloatWidget
-    case "expandable_text":
-      return ExandableTextWidget
-    case "select":
-      return SelectWidget
-    case "llm_provider_model":
-      return LlmProviderWidget
-    case "history":
-      return HistoryTypeWidget
-    case "keywords":
-      return KeywordsWidget
-    default:
-      return DefaultWidget
-  }
-}
-
-interface WidgetParams {
-  nodeId: string;
-  name: string;
-  label: string;
-  helpText: string;
-  paramValue: string | string[];
-  inputError: string | undefined;
-  updateParamValue: (event: React.ChangeEvent<HTMLTextAreaElement | HTMLSelectElement | HTMLInputElement>) => any;
-  schema: PropertySchema
-  nodeParams: NodeParams
-  required: boolean
-}
-
-function DefaultWidget(props: WidgetParams) {
-  return (
-    <InputField label={props.label} help_text={props.helpText} inputError={props.inputError}>
-      <input
-        className="input input-bordered w-full"
-        name={props.name}
-        onChange={props.updateParamValue}
-        value={props.paramValue}
-        type="text"
-        required={props.required}
-      ></input>
-    </InputField>
-  );
-}
-
-function FloatWidget(props: WidgetParams) {
-  return <InputField label={props.label} help_text={props.helpText} inputError={props.inputError}>
-    <input
-      className="input input-bordered w-full"
-      name={props.name}
-      onChange={props.updateParamValue}
-      value={props.paramValue}
-      type="number"
-      step=".1"
-      required={props.required}
-    ></input>
-  </InputField>
-}
-
-function ToggleWidget(props: WidgetParams) {
-  const onChangeCallback = (event: React.ChangeEvent<HTMLInputElement>) => {
-    event.target.value = event.target.checked ? "true" : "false";
-    props.updateParamValue(event);
-  };
-  return (
-    <InputField label={props.label} help_text={props.helpText} inputError={props.inputError}>
-      <input
-        className="toggle"
-        name={props.name}
-        onChange={onChangeCallback}
-        checked={props.paramValue === "true"}
-        type="checkbox"
-      ></input>
-    </InputField>
-  );
-}
-
-function SelectWidget(props: WidgetParams) {
-  const options = getSelectOptions(props.schema);
-  return <InputField label={props.label} help_text={props.helpText} inputError={props.inputError}>
-    <select
-      className="select select-bordered w-full"
-      name={props.name}
-      onChange={props.updateParamValue}
-      value={props.paramValue}
-      required={props.required}
-    >
-      {options.map((option) => (
-        <option key={option.value} value={option.value}>
-          {option.label}
-        </option>
-      ))}
-    </select>
-  </InputField>
-}
-
-function LlmProviderWidget(props: WidgetParams) {
-  return (
-    <InputField label={props.label} help_text={props.helpText} inputError={props.inputError}>
-      <LlmWidget
-        name={props.name}
-        nodeId={props.nodeId}
-        providerId={concatenate(props.nodeParams.llm_provider_id)}
-        providerModelId={concatenate(props.nodeParams.llm_provider_model_id)}
-      ></LlmWidget>
-    </InputField>
-  );
-}
-
-function ExandableTextWidget(props: WidgetParams) {
-  return (
-    <ExpandableTextWidget
-      humanName={props.label}
-      name={props.name}
-      onChange={props.updateParamValue}
-      value={props.paramValue}
-      help_text={props.helpText}
-      inputError={props.inputError}>
-    </ExpandableTextWidget>
-  );
-}
-
-function KeywordsWidget(props: WidgetParams) {
-  return <KeywordsWidgetInner nodeId={props.nodeId} params={props.nodeParams} inputError={props.inputError}/>
-}
-
-
-function HistoryTypeWidget(props: WidgetParams) {
-  return (
-    <HistoryTypeWidgetInner
-      onChange={props.updateParamValue}
-      name={props.name}
-      schema={props.schema}
-      historyType={concatenate(props.paramValue)}
-      historyName={concatenate(props.nodeParams["history_name"])}
-      help_text={props.helpText}
-    ></HistoryTypeWidgetInner>
-  );
-}
